==> 04_clustering.Rmd <==
---
title: "Dimensionality reduction & Clustering"
output:
  html_notebook:
    toc: true
    toc_float: true
    toc_depth: 2
---


```{r}
source("src/00_library.R")
seu <- readRDS(paste0(dirs["rds"],"seu_corrected.RDS"))
```


```{r}
seu <- RunPCA(seu, features = VariableFeatures(object = seu))
seu <- RunUMAP(seu, dims = 1:seurat[["ndims_use"]])
```


```{r}
seu <- FindNeighbors(seu, dims = 1:seurat[["ndims_use"]])
seu <- FindClusters(seu, resolution = c(seq(0, 2, 0.2)))
```


# Check cluster resolution
```{r}
# Check amount of clusters
sapply(grep("res",colnames(seu@meta.data),value = TRUE),
       function(x) length(unique(seu@meta.data[,x])))
clustree(seu, prefix = "integrated_snn_res.")
```

```{r}
# Check how clusters correspond with UMAP
for(res in grep("res", colnames(seu@meta.data), value = TRUE)) {
  print(DimPlot(seu, reduction = "umap", group.by = res, label = TRUE) +
          ggtitle(res))
}
```


```{r}
# Determine best resolution, save in settings, and run clustering again
seu <- FindClusters(seu, resolution = seurat[["resolution_corrected"]])
# Cluster number + 1
seu$seurat_clusters <- as.factor(as.integer(seu$seurat_clusters))
Idents(seu) <- seu$seurat_clusters
# Plot Dimplot
DimPlot(seu, reduction = "umap", group.by = "seurat_clusters", label = TRUE) +
  ggtitle(paste("Final resolution:", seurat[["resolution_corrected"]]))
```


# Marker genes per cluster
```{r}
# Switch back to RNA as default data slot
DefaultAssay(seu) <- "RNA"
# Check marker genes (have at least 20 per cluster)
markers <- FindAllMarkers(object = seu,
                          only.pos = TRUE,
                          min.pct = 0.25,
                          logfc.threshold = 0.25,
                          test.use = "MAST")
markers <- markers[markers$p_val_adj < 0.05,]

# Save markers as csv
write.csv(markers,
          paste0(dirs[["output"]],"cluster_markers_corrected.csv"),
          row.names = FALSE)
```

# Remove contaminating clusters
```{r}
# Cluster 10 + 13: Express preDC markers
# Cluster 15: Express monocyte markers
seu <- subset(seu, idents = c(10, 13, 15), invert = TRUE)
```

# Redo dimensionality reduction and clustering
```{r}
# Switch back to integrated assay
DefaultAssay(seu) <- "integrated"
# RunPCA, UMAP, FindNeighbors and compute clustering
seu <- RunPCA(seu, features = VariableFeatures(object = seu))
seu <- RunUMAP(seu, dims = 1:seurat[["ndims_use"]])
seu <- FindNeighbors(seu, dims = 1:seurat[["ndims_use"]])
seu <- FindClusters(seu, resolution = seurat[["resolution_corrected"]])
# Cluster number + 1
seu$seurat_clusters <- as.factor(as.integer(seu$seurat_clusters))
Idents(seu) <- seu$seurat_clusters
# Plot Dimplot
DimPlot(seu, reduction = "umap", group.by = "seurat_clusters", label = TRUE) +
  ggtitle(paste("Final resolution:", seurat[["resolution_corrected"]]))
```

# Merge highly similar clusters (3+6 and 7+13)
```{r}
new_ids <- c(1, 2, 3, 4, 5, 3, 6, 7, 8, 9, 10, 11, 6, 12)
names(new_ids) <- levels(seu)
seu <- RenameIdents(seu, new_ids)
seu$seurat_clusters <- Idents(seu)
DimPlot(seu, reduction = "umap", label = TRUE) +
  ggtitle("New clusters after merging")
```


# Rerun marker genes for new clusters
```{r}
# Switch back to RNA assay
DefaultAssay(seu) <- "RNA"
# Check marker genes (have at least 20 per cluster)
markers <- FindAllMarkers(object = seu,
                          only.pos = TRUE,
                          min.pct = 0.25,
                          logfc.threshold = 0.25,
                          test.use = "MAST")
markers <- markers[markers$p_val_adj < 0.05,]

# Save markers as csv
write.csv(markers,
          paste0(dirs[["output"]],"cluster_markers_corrected_clean.csv"),
          row.names = FALSE)
```

# Gene ontology enrichment dotplot
```{r}
goMtx <- data.frame()
goMtxTop <- data.frame()

for(cluster in unique(markers$cluster)) {
  currentMtx <- enrich_enrichr(markers[which(markers$cluster==cluster),"gene"],
                      paste0("GOBP21_cluster_", cluster),
                      "GO_Biological_Process_2021",
                      "output/donor_corrected/DEG/")
  currentMtx$cluster <- cluster
  goMtx <- rbind(goMtx, currentMtx)
  goMtxTop <- rbind(goMtxTop, currentMtx[1:5,])
}

goMtx <- goMtx[goMtx$Adjusted.P.value<0.05,]

# Retrieve GOs from all comparisons from top table
goMtxFinal <- goMtx[which(goMtx$Term %in% goMtxTop$Term),]

# Remove(GO:XXXXX) from Term (it's too long)
goMtxFinal$Term <- sub("\\ \\(.*", "", goMtxFinal$Term)

# Plot
ggplot(goMtxFinal, aes(size=-log10(Adjusted.P.value),
                       x=factor(cluster),
                       y=Term,
                       color=factor(cluster))) +
  geom_point() +
  theme_classic() +
  scale_color_manual(values = colors[["seurat_clusters"]][1:12]) +
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1))
```



```{r}
saveRDS(seu, paste0(dirs["rds"],"seu_corrected_clean.RDS"))
```
